adjusted imports, gitignore, refactor into setup.go, add setup.go
