+++ taco/.gitignore
@@ -1,3 +1,3 @@
-taco
-statesman
-terraform-provider-opentaco
+/taco
+/statesman
+/terraform-provider-opentaco
